+++ lib/src/storage.dart
@@ -102,7 +102,10 @@ class StorageReference
 
   Future delete() => handleThenable(jsObject.delete());
 
-  Future<String> getDownloadURL() => handleThenable(jsObject.getDownloadURL());
+  Future<Uri> getDownloadURL() async {
+    var uriString = await handleThenable(jsObject.getDownloadURL());
+    return Uri.parse(uriString);
+  }
 
   Future<FullMetadata> getMetadata() => handleThenableWithMapper(
       jsObject.getMetadata(), (m) => new FullMetadata.fromJsObject(m));
@@ -143,7 +146,7 @@ class FullMetadata
     extends _UploadMetadataBase<storage_interop.FullMetadataJsImpl> {
   String get bucket => jsObject.bucket;
 
-  List<String> get downloadURLs => new List.from(jsObject.downloadURLs);
+  List<Uri> get downloadURLs => jsObject.downloadURLs.map(Uri.parse).toList();
 
   String get fullPath => jsObject.fullPath;
 
@@ -155,9 +158,9 @@ class FullMetadata
 
   int get size => jsObject.size;
 
-  String get timeCreated => jsObject.timeCreated;
+  DateTime get timeCreated => DateTime.parse(jsObject.timeCreated);
 
-  String get updated => jsObject.updated;
+  DateTime get updated => DateTime.parse(jsObject.updated);
 
   factory FullMetadata(
           {String bucket,
@@ -201,7 +204,7 @@ class UploadMetadata
           String contentEncoding,
           String contentLanguage,
           String contentType,
-          Map customMetadata}) =>
+          Map<String, String> customMetadata}) =>
       new UploadMetadata.fromJsObject(new storage_interop.UploadMetadataJsImpl(
           md5Hash: md5Hash,
           cacheControl: cacheControl,
@@ -302,7 +305,7 @@ class UploadTaskSnapshot
     extends JsObjectWrapper<storage_interop.UploadTaskSnapshotJsImpl> {
   int get bytesTransferred => jsObject.bytesTransferred;
 
-  String get downloadURL => jsObject.downloadURL;
+  Uri get downloadURL => Uri.parse(jsObject.downloadURL);
 
   FullMetadata _metadata;
   FullMetadata get metadata {
@@ -341,7 +344,7 @@ class UploadTaskSnapshot
       case "error":
         return TaskState.ERROR;
       default:
-        return null;
+        throw 'Unknown state "${jsObject.state}" please file a bug.';
     }
   }
 
@@ -416,8 +419,9 @@ abstract class _SettableMetadataBase<
     jsObject.contentType = s;
   }
 
-  Map get customMetadata => dartify(jsObject.customMetadata);
-  void set customMetadata(Map m) {
+  Map<String, String> get customMetadata =>
+      dartify(jsObject.customMetadata) as Map<String, String>;
+  void set customMetadata(Map<String, String> m) {
     jsObject.customMetadata = jsify(m);
   }