+++ test/test_util.dart
@@ -1,0 +1,4 @@
+import 'package:path/path.dart' as p;
+
+String validDatePath() =>
+    p.join('pkg_firebase3_test', new DateTime.now().toUtc().toIso8601String());